Completing logic for add to cart product with livewire.

+++ app/Http/Livewire/Products/ProductShow.php
@@ -11,11 +11,20 @@ class ProductShow extends Component
 
     public function render()
     {
-        return view('livewire.products.product-show', ['products' => Product::query()->take(4)->get()]);
+        $cart = \Cart::get($this->product->id);
+        $this->product->in_cart = !empty($cart);
+
+        $in_wishlist = auth()->user()->wishlist()->where('product_id', $this->product->id)->first();
+        $this->product->in_wishlist = !empty($in_wishlist);
+
+        return view('livewire.products.product-show', [
+            'product' => $this->product,
+            'products' => Product::query()->take(4)->get()
+        ]);
     }
 
-    public function addToCart(Product $product) {
-
+    public function addToCart(Product $product)
+    {
         \Cart::add([
             'id' => $product->id, // inique row ID
             'name' => $product->name,
@@ -29,37 +38,33 @@ class ProductShow extends Component
             ],
         ]);
 
+        $this->emit('refreshCart');
         session()->flash('message', 'Product has been added in cart!');
-        return redirect()->route('products.show',['slug' => $product->slug]);
-
+        return $this->render();
     }
 
     public function removeFromCart(Product $product)
     {
         \Cart::remove($product->id);
 
+        $this->emit('refreshCart');
         session()->flash('message', 'Product has been removed from cart!');
-        return redirect()->route('products.show',['slug' => $product->slug]);
-
+        return $this->render();
     }
 
-    public function addToWishList(Product $product) {
-
-        $user = auth()->user();
-        $user->wishlist()->create(['product_id' => $product->id]);
+    public function addToWishList(Product $product)
+    {
+         auth()->user()->wishlist()->create(['product_id' => (int)$product->id]);
 
         session()->flash('message', 'Product has been added in Wish List!');
-        return redirect()->route('products.show',['slug' => $product->slug]);
-
+        return $this->render();
     }
 
-    public function removeFromWishList(Product $product) {
-
-        $user = auth()->user();
-        $user->wishlist()->where('product_id' , $product->id)->first()->delete();
+    public function removeFromWishList(Product $product)
+    {
+        auth()->user()->wishlist()->where('product_id', $product->id)->first()->delete();
 
         session()->flash('message', 'Product has been removed from Wish List!');
-        return redirect()->route('products.show',['slug' => $product->slug]);
-
+        return $this->render();
     }
 }
